examples/AIcargo/requirements.txt
Django~=3.2
protobuf~=3.15.8
loguru~=0.5.3
tensorflow-serving-api~=2.4.1
cryptography~=2.8
djangorestframework~=3.12.4
django-cors-headers~=3.7.0
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0
zipfile36~=0.1.3
Pillow~=8.2.0
psycopg2~=2.8.6
gunicorn~=20.1.0
# sentry-sdk~=1.1.0
gevent~=21.1.2